requirements/base.txt
Django==3.2
python-environ
django-cors-headers
PyJWT
python-environ
pre-commit